refactor: apply ddd and cqrs architecture

+++ Domain/Shared/Repositories/IBaseRepository.cs
@@ -4,7 +4,7 @@ public interface IBaseRepository<TEntity> where TEntity:  class
 {
     Task AddAsync(TEntity entity);
     Task <IEnumerable<TEntity>> GetAllAsync();
-    Task<TEntity?> GetByIdAsync(long id);
+    Task<TEntity?> GetByIdAsync(int id);
     void Update(TEntity entity);
     void Delete(TEntity entity);
 }
